adds jotai
adds jotai and some state management for control panel
