+++ components/GeoJsonMap.tsx
@@ -1,5 +1,7 @@
 import * as React from 'react';
 import Map, { Layer, Source, FillLayer } from 'react-map-gl';
+import { useAtom } from 'jotai';
+import { mapAtom } from '@/lib/mapStore';
 import { mapboxAccessToken } from '@/pages/index'
 
   const jsonFillLayer: FillLayer = {
@@ -12,8 +14,10 @@ import { mapboxAccessToken } from '@/pages/index'
   };
 
 function GeoJsonMap({data}:any) {
+  const [, setMap] = useAtom(mapAtom);
   return (
     <Map
+      ref={(ref) => setMap(ref)}
       initialViewState={{
         longitude: -100,
         latitude: 40,
