+++ package.json
@@ -16,6 +16,7 @@
     "eslint": "8.31.0",
     "eslint-config-next": "13.1.1",
     "immutable": "^4.2.1",
+    "jotai": "^1.13.0",
     "mapbox-gl": "^2.12.0",
     "next": "13.1.1",
     "react": "18.2.0",
